% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/api.R
\name{search}
\alias{search}
\title{Search QPX for flights}
\usage{
search(origin, dest, startDate, returnDate, adultCount = 1, ...)
}
\arguments{
\item{origin}{Departing Airport Code}

\item{dest}{Destination Airport Code}

\item{startDate}{Departure date (yyyy-mm-dd)}

\item{startDate}{Return date (yyyy-mm-dd)}

\item{numPassengers}{Number of people traveling}
}
\description{
Search QPX for flights
}

